Complete raster optimisation code
I have also split out CutPlan objects into a separate file to reduce the length of planner.py.
